requirements.txt
flask
python-dotenv
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
flask-bootstrap
flask-mail
pyjwt
email_validator